Integration Tests added to Tenant + added Integration test files seperate from GETTER/SETTER/CONSTRUCTOR test files
